requirements.txt
werkzeug
Flask
Flask-SQLAlchemy
requests
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
